Terminando de modificar todo para la jerarquia de CIA-CCenter

+++ sigo/db/migrate/20140403152449_remove_code_column_from_specifics.rb
@@ -1,5 +1,5 @@
 class RemoveCodeColumnFromSpecifics < ActiveRecord::Migration
   def change
-  	remove_column :specifics, :code, :integer
+  	remove_column :specifics, :code, :integer if table_exists? :specifics
   end
 end
